+++ src/main/java/net/jasper/mod/gui/RecordingSelector.java
@@ -3,6 +3,7 @@ package net.jasper.mod.gui;
 
 import net.jasper.mod.PlayerAutomaClient;
 import net.jasper.mod.automation.PlayerRecorder;
+import net.jasper.mod.util.PlayerController;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.Screen;
@@ -92,6 +93,12 @@ public class RecordingSelector extends Screen {
     private void onDelete() {
         RecordingSelectionListWidget.RecordingEntry recEntry = this.recordingSelectionList.getSelectedOrNull();
         if (recEntry != null) {
+            boolean deleteSuccess = recEntry.file.delete();
+            if (!deleteSuccess) {
+                PlayerAutomaClient.LOGGER.warn("Could not delete recording file " + recEntry.fileName);
+                this.close();
+                PlayerController.writeToChat(Text.translatable("playerautoma.messages.deleteFailed"));
+            }
             this.recordingSelectionList.updateFiles();
         }
     }