+++ src/main/java/net/jasper/mod/util/PlayerController.java
@@ -33,9 +33,14 @@ public class PlayerController {
     }
 
     public static void writeToChat(String message) {
-        if (MinecraftClient.getInstance().player != null) {
+        // Only if in-game and enabled options
+        if (MinecraftClient.getInstance().player != null && PlayerAutomaOptionsScreen.writeStateToChatOption.getValue()) {
             MinecraftClient.getInstance().player.sendMessage(Text.of(message));
         }
+    }
+
+    public static void writeToChat(Text message) {
+        writeToChat(message.getString());
     }
 
     public static void clickSlot(SlotClick click) {